app/requirements.txt
certifi==2024.8.30
elastic-transport==8.15.0
elasticsearch==8.15.1
mysql-connector==2.2.9
python-dotenv==1.0.1
urllib3==2.2.3

pip~=23.2.1